+++ server/src/modules/groupMessage.module.js
@@ -5,13 +5,11 @@ const groupMessageSchema = new Schema(
     groupId: { type: Schema.Types.ObjectId, ref: "Group", required: true },
     sender: { type: Schema.Types.ObjectId, ref: "User", required: true },
     message: { type: String },
-    image: {
-      type: String,
-    },
+    image: { type: String },
   },
   { timestamps: true }
 );
 
-const groupMessage =  model("GroupMessage", groupMessageSchema);
+const groupMessage = model("GroupMessage", groupMessageSchema);
 
 export default groupMessage;