+++ server/src/routes/groupRoute.route.js
@@ -1,9 +1,11 @@
 import express from "express";
-import { createGroup } from "../controllers/group.controller.js";
+import { createGroup, getGroup } from "../controllers/group.controller.js";
 import { authenticate } from "../middleware/auth.middleware.js";
 
 const route = express.Router();
 
 route.post("/createGroup", authenticate, createGroup);
+
+route.get("/getGroups/:id", authenticate, getGroup);
 
 export default route;